Phase K Wave 11: 0x0012 TURN + 0x0013 TILT survival round-trip + wire-relevant long sweep
* New test SectorTurnTiltTests.TurnAndTilt_DoNotBreakConnection_RequestTimeStillRoundTrips —
  sends canonical 8-byte TURN payload {int32 GameID; float Intensity} (GameID=0, intensity=0.5f),
  then TILT (intensity=-0.25f), then 0x0044 REQUEST_TIME with a sentinel tick, asserts the 0x0034
  CLIENT_SET_TIME echo. Survival-probe shape because Moveable::Turn/Tilt fan-out via
  SendPositionalUpdate goes to OTHER observers in the visibility list, never back to the
  originator — direct reply assertion fundamentally unreachable for a single-player test.

* Server fix: HandleTurn / HandleTilt local PacketTurn struct
  (server/src/PlayerConnection.cpp:1793-1827) was { long GameID; float Intensity } — 8B on
  Win32 (sizeof(long)==4) matching wire exactly, but 12B on Linux x86_64. Intensity read
  from offset 8, past the end of the 8B wire payload into undefined memory data[8..],
  potentially propagating NaN through Moveable::Turn(intensity) on the next physics tick.
  Narrowed GameID to int32_t.

* Wire-relevant long sweep (the user 'fix all longs' pass after Waves 7/9 kept hitting the
  same sizeof(long) class): pushed back on literal global s/long/int32_t/ — internal
  counters, loop indices, time deltas, queue API parameters (CheckQueue's
  'long *player_id') should stay 64-bit on Linux because they don't touch the wire.
  Committed to the substantively-correct version: sweep every long that touches the wire.

  * server/src/SaveManager.cpp ~14 sites — every long X = *((long *) &data[N]) inbound
    wire read across the save-state opcodes (HandelInfraction, HandleNewRecipe,
    HandleManufactureAttempt, HandleAdvanceLevel, HandleChangeInventory,
    HandleChangeEquipment, HandleChangeAmmo, HandleStorePosition, HandleAdvanceMission,
    HandleAdvanceMissionFlags, HandleMissionRemove, HandleMissionComplete,
    HandleDiscoverNav, HandleExploreNav, HandleSetSkillPoints,
    HandleSetRegisteredStarbase, HandleNewWarnLevel, HandleChangeFieldRespawn).

  * server/src/PlayerConnection.cpp ~9 sites beyond the Wave 11 PacketTurn fix:
    SendOpcode wire writes (792/796 — *((long *) &buffer[4]) = GameID() was writing 8B
    over a 4B slot, corrupting the next 4B), wire read 873, line 1310 AdvancedPositional-
    Update slot stride (must be 4B — confirmed by length = 2 + 4 * index), 1679 size
    header, 3649 TradeAction GameID into a 5B buffer, 10191 chat-stream GameID.

  * server/src/PlayerConnection.cpp GalaxyMap struct (10669-10674) —
    { long Type; long Size; long PlayerID; char Variable[64]; long unknown; } (96B on
    Linux) → all int32_t (80B matching the Win32 wire shape 4+4+4+64+4).

  * proxy/ClientToSectorServer.cpp 4 sites — inbound wire reads on the
    sector dispatch path.

  * proxy/Connection.cpp + login-server/Net7SSL/Connection.cpp 5 sites each — the RC4
    key-exchange handshake length prefix harmonised to the canonical 4B big-endian width
    (matches CLI client RsaHandshake.cs:101). Linux↔Linux worked by accident before —
    ntohl truncates and the LE zero-init of the high bytes made the 8B read+write
    effectively 4B.

  * proxy/ClientToGlobalServer.cpp + login-server/Net7SSL/ClientToGlobalServer.cpp
    (each 2 sites) — HandleDeleteCharacter + HandleGlobalTicketRequest wire reads. Same
    class as the Wave 5 PacketMethods.h::ExtractLong fix.

* CLAUDE.md server-integrity compliance: every fix restores a Win32 wire width Linux
  was over-reading or over-writing by 4B. No server permissiveness was added; the test
  intensity values (0.5f, -0.25f) are within the retail client's normal stick range.

* Pool extended: cli_test08 (id 9000008) added to seed.sql + TestAccounts.Pool[7] —
  dedicated to SectorTurnTiltTests so Create/Delete doesn't collide with Pool[3..6]
  owned by SectorLogin/SectorChat/SectorRequestTime/SectorStartAck.

* Coverage 20/207 → 22/207 (10.6%). MinTestedCount 20→22. +0x0012 TURN, +0x0013 TILT.

Full integration suite 43/43 green from a fresh docker compose down -v rebuild.

+++ login-server/Net7SSL/ClientToGlobalServer.cpp
@@ -187,7 +187,11 @@ void Connection_B::HandleGlobalConnect()
 
 void Connection_B::HandleDeleteCharacter()
 {
-    long character_slot = ntohl(*(long *) m_RecvBuffer);
+    // Phase K Wave 11: wire field is 4B big-endian (Win32 sizeof(long)==4).
+    // `long *` on Linux x86_64 reads 8B; ntohl truncates to uint32_t so it
+    // works by accident on little-endian, but is asymmetric with the
+    // canonical wire spec.
+    long character_slot = ntohl(*(uint32_t *) m_RecvBuffer);
 	long avatar_id = g_AccountMgr->GetAvatarID(m_AccountUsername, character_slot);
 
 	LogMessage("Delete character %d\n", avatar_id);
@@ -226,7 +230,8 @@ void Connection_B::SendAvatarList(long account_id)
 void Connection_B::HandleGlobalTicketRequest()
 {
     // The player selected a character
-    long char_slot = ntohl(*(long *) m_RecvBuffer);
+    // Phase K Wave 11: wire field is 4B; see HandleDeleteCharacter.
+    long char_slot = ntohl(*(uint32_t *) m_RecvBuffer);
     long avatar_id = g_AccountMgr->GetAvatarID(m_AccountUsername, char_slot);
 
 	m_Mutex.Lock();
